requirements.txt
# Core data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.25.0

# Kronos (install from GitHub — see README)
# git+https://github.com/shiyu-coder/Kronos.git

# HuggingFace
torch>=2.0.0
transformers>=4.40.0
huggingface_hub>=0.20.0
einops>=0.7.0

# Sentiment
# FinBERT: ProsusAI/finbert via transformers

# Technical analysis
ta>=0.10.2

# Reddit
praw>=7.7.0

# Storage
sqlalchemy>=2.0.0

# Dashboard
streamlit>=1.35.0
plotly>=5.20.0

# Utilities
python-dotenv>=1.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
anthropic>=0.25.0
